services/web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
selenium
PyJWT
apscheduler
Flask-Migrate